+++ packages/app/src/features/my-portfolio/components/my-wallet-panel/MyWalletPanel.tsx
@@ -1,6 +1,6 @@
 import { TokenWithBalance } from '@/domain/common/types'
-import { Switch } from '@/ui/atoms/new/switch/Switch'
 import { Panel } from '@/ui/atoms/panel/Panel'
+import { Switch } from '@/ui/atoms/switch/Switch'
 import { Info } from '@/ui/molecules/info/Info'
 import { MyWalletChart } from './MyWalletChart'
