+++ packages/app/src/ui/molecules/labeled-switch/LabeledSwitch.tsx
@@ -1,5 +1,5 @@
 import { Label } from '@/ui/atoms/label/Label'
-import { Switch } from '@/ui/atoms/new/switch/Switch'
+import { Switch } from '@/ui/atoms/switch/Switch'
 import { ComponentProps, useId } from 'react'
 
 interface LabeledSwitchProps extends ComponentProps<typeof Switch> {
